+++ crdesigner/ui/gui/mwindow/toolboxes/obstacle_toolbox/obstacle_toolbox_ui.py
@@ -395,9 +395,6 @@ class ObstacleToolboxUI(Toolbox):
         if self.obstacle_dyn_stat.currentText() == "Dynamic":
             self.toggle_dynamic_static()
 
-        if self.obstacle_dyn_stat.currentText() == "Dynamic":
-            self.remove_dynamic_fields()
-            self.toggle_dynamic_static()
 
     def adjust_obstacle_type_dropdown(self, selected_type):
         self.obstacle_type.clear()