+++ tests/map_conversion/test_opendrive_conversion.py
@@ -14,7 +14,7 @@ from tests.map_conversion.utils import elements_equal
 __author__ = "Benjamin Orthen, Sebastian Maierhofer"
 __copyright__ = "TUM Cyber-Physical Systems Group"
 __credits__ = ["Priority Program SPP 1835 Cooperative Interacting Automobiles, BMW Car@TUM"]
-__version__ = "0.5"
+__version__ = "0.5.1"
 __maintainer__ = "Sebastian Maierhofer"
 __email__ = "commonroad@lists.lrz.de"
 __status__ = "Released"
@@ -59,12 +59,12 @@ class TestOpenDriveToCommonRoadConversion(unittest.TestCase):
             # set same date so this won't change the comparison
             date = time.strftime("%Y-%m-%d", time.localtime())
             tree_import.set("date", date)
-            writer.root_node.set("date", date)
+            writer._file_writer.root_node.set("date", date)
 
             generate_unique_id(0)  # reset ID counter for next test case
 
             # compare both element trees
-            return elements_equal(tree_import, writer.root_node)
+            return elements_equal(tree_import, writer._file_writer.root_node)
 
     def test_basic_opendrive(self):
         """Basic test with a junction in the middle."""